requirements.txt
langchain>=0.3
langchain-openai>=0.3
langchain-mcp-adapters>=0.1
langgraph>=0.4
python-dotenv>=1.0
